switched to openai api instead of hf

+++ audio_scraper/transcriber.py
@@ -1,6 +1,13 @@
 from pathlib import Path
-import librosa
-from transformers import WhisperProcessor, WhisperForConditionalGeneration
+from typing import Literal
+import warnings
+
+from loguru import logger as log
+from tqdm import tqdm
+import whisper
+
+warnings.filterwarnings("ignore")
+MODEL = Literal["tiny", "base", "small", "medium", "large"]
 
 
 class Transcriber:
@@ -9,27 +16,44 @@ class Transcriber:
     """
 
     transcript_dir: Path
+    model: whisper.Whisper
 
-    def __init__(self, data_dir: Path):
+    def __init__(self, data_dir: Path, model_size: MODEL):
         self.transcript_dir = data_dir / "transcripts"
         self.transcript_dir.mkdir(exist_ok=True, parents=True)
 
-        self.processor = WhisperProcessor.from_pretrained("openai/whisper-tiny")
-        self.model = WhisperForConditionalGeneration.from_pretrained("openai/whisper-tiny")
+        self.model = whisper.load_model(model_size)
 
-    def __call__(self, filepath: str, language: str = "en") -> str:
-        wav, _ = librosa.load(filepath, sr=16_000)
-        input_features = self.processor(wav, return_tensors="pt", sampling_rate=16_000).input_features 
-        forced_decoder_ids = self.processor.get_decoder_prompt_ids(language=language, task="translate")
-        predicted_ids = self.model.generate(input_features, forced_decoder_ids=forced_decoder_ids)
-        transcript = self.processor.batch_decode(predicted_ids, skip_special_tokens=True)[0]
-
+    def __call__(self, filepath: str, skip_existing: bool = False) -> str:
         basename = Path(filepath).parent.name
         download_dir = self.transcript_dir / basename
         download_dir.mkdir(exist_ok=True)
 
         filename = Path(filepath).name.split(".")[0]
         result_path = download_dir / f"{filename}.txt"
+
+        if result_path.exists() and skip_existing:
+            log.warning(f"{result_path} exists. Skipping.")
+            return result_path.read_text()
+
+        result = self.model.transcribe(filepath)
+        transcript = result["text"]
         result_path.write_text(transcript)
 
         return transcript
+
+
+def transcribe():
+    import argparse
+    parser = argparse.ArgumentParser()
+    parser.add_argument("data_dir", type=Path)
+    parser.add_argument("--model", choices=MODEL.__args__, default="tiny")
+    parser.add_argument("--skip-existing", action="store_true")
+    args = parser.parse_args()
+
+    transcriber = Transcriber(args.data_dir, args.model)
+
+    audio_files = list((args.data_dir / "chunks").rglob("*"))
+    audio_files = list(filter(lambda f: f.is_file(), audio_files))
+    for af in tqdm(audio_files):
+        transcriber(str(af), args.skip_existing)
